requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings>=2.2.0

google-generativeai
gspread
google-auth
openpyxl